[Enh]: Added Quarto and friends.
Added a category Notebooks, moved Jupyter to it and created Quarto here.
Both serve as small bundles that would be a good candidates for an options in future layer DataScience.
Something like: `DataScience +notebook=quarto`

Quarto has jukit, img-clip, magma, slime and otter for better REPL like behaviour. As far as I understand, they're not dependencies but
just compliments so it would be better to atomize things a little bit further.